Countersinking
I bought 7/16” screws so I’m adding a way to add a countersink (need to
check that I’m correct about that) so that I can get away with more
thickness for the same amount of screw.

I printed today and I learned a lot, including not just the above but
also that I need DESPERATELY to tune my printer and that this works
well with a brim.